fix: The provided autocmd was not actually saving session on write
This fixes it. It also contains an aditional fix: Using BufWritePre will allow users running notifications on BufWritePost in nvim. This is necessary because running `save_current_session()` will clear all notifications. Also we are writing 2kb when we create a session, so the price of using BufWritePre is negligible.